+++ src/main/java/forumdb/DAO/ServiceDAO.java
@@ -18,10 +18,10 @@ public class ServiceDAO {
 
     public InfoAboutDB getStatus() throws DataAccessException {
         final InfoAboutDB currentStatus = new InfoAboutDB();
-        currentStatus.setForum(jdbcTemplate.queryForObject("SELECT count(*) FROM Forum;", Integer.class));
-        currentStatus.setPost(jdbcTemplate.queryForObject("SELECT count(*) FROM Post;", Integer.class));
-        currentStatus.setThread(jdbcTemplate.queryForObject("SELECT count(*) FROM Thread;", Integer.class));
-        currentStatus.setUser(jdbcTemplate.queryForObject("SELECT count(*) FROM \"User\";", Integer.class));
+        currentStatus.setForum(jdbcTemplate.queryForObject("SELECT count(*) FROM Forum;", Long.class));
+        currentStatus.setPost(jdbcTemplate.queryForObject("SELECT count(*) FROM Post;", Long.class));
+        currentStatus.setThread(jdbcTemplate.queryForObject("SELECT count(*) FROM Thread;", Long.class));
+        currentStatus.setUser(jdbcTemplate.queryForObject("SELECT count(*) FROM \"User\";", Long.class));
 
         return currentStatus;
     }
