+++ src/main/java/forumdb/Model/Post.java
@@ -7,19 +7,19 @@ public class Post {
     private String author;
     private String created;
     private String forum;
-    private Integer id;
+    private Long id;
     private Boolean isEdited;
     private String message;
-    private Integer parent;
-    private Integer thread;
-    private Integer forum_id;
+    private Long parent;
+    private Long thread;
+    private Long forum_id;
     private Object[] path;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public Integer getParent() {
+    public Long getParent() {
         return parent;
     }
 
@@ -39,7 +39,7 @@ public class Post {
         return forum;
     }
 
-    public Integer getThread() {
+    public Long getThread() {
         return thread;
     }
 
@@ -47,7 +47,7 @@ public class Post {
         return created;
     }
 
-    public Integer getForumID() {
+    public Long getForumID() {
         return forum_id;
     }
 
@@ -60,11 +60,11 @@ public class Post {
         this.message = message;
     }
 
-    public void setParent(Integer parent) {
+    public void setParent(Long parent) {
         this.parent = parent;
     }
 
-    public void setId(Integer id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -76,7 +76,7 @@ public class Post {
         this.forum = forum;
     }
 
-    public void setThread(Integer thread) {
+    public void setThread(Long thread) {
         this.thread = thread;
     }
 
@@ -84,7 +84,7 @@ public class Post {
         this.created = created.toInstant().toString();
     }
 
-    public void setForumID(Integer forum_id) {
+    public void setForumID(Long forum_id) {
         this.forum_id = forum_id;
     }
