+++ src/main/java/forumdb/Model/Thread.java
@@ -7,18 +7,18 @@ public class Thread {
     private String author;
     private String created;
     private String forum;
-    private Integer id;
+    private Long id;
     private String message;
     private String title;
     private String slug;
-    private Integer votes;
-    private Integer forum_id;
+    private Long votes;
+    private Long forum_id;
 
-    public Integer getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
@@ -54,11 +54,11 @@ public class Thread {
         this.message = message;
     }
 
-    public int getVotes() {
+    public Long getVotes() {
         return votes;
     }
 
-    public void setVotes(int votes) {
+    public void setVotes(Long votes) {
         this.votes = votes;
     }
 
@@ -78,11 +78,11 @@ public class Thread {
         this.created = created.toInstant().toString();
     }
 
-    public Integer getForumID() {
+    public Long getForumID() {
         return forum_id;
     }
 
-    public void setForumID(Integer forum_id) {
+    public void setForumID(Long forum_id) {
         this.forum_id = forum_id;
     }
 }
